refactor(tools_): split measurement_info into a package
Break measurement_info_.py into a measurement_info/ package with one
MeasurementInfo subclass per module, and rename the class attribute
_measurement_info_class[es] -> _measurement_infoclass[es] on every
operation. Docs rst file renamed to match. Pure mechanical migration —
no behavior changes.
