+++ docs/source/api_reference/phenotypic.tools_.measurement_info.rst
@@ -1,0 +1,7 @@
+phenotypic.tools\_.measurement\_info package
+============================================
+
+.. automodule:: phenotypic.tools_.measurement_info
+   :members:
+   :show-inheritance:
+   :undoc-members:
