--- docs/source/api_reference/phenotypic.tools_.measurement_info_.rst
@@ -1,7 +1,0 @@
-phenotypic.tools\_.measurement\_info\_ module
-=============================================
-
-.. automodule:: phenotypic.tools_.measurement_info_
-   :members:
-   :show-inheritance:
-   :undoc-members:
