+++ src/phenotypic/grid/_auto_grid_finder.py
@@ -10,7 +10,7 @@ import pandas as pd
 import numpy as np
 
 from phenotypic.abc_ import GridFinder
-from phenotypic.tools_.measurement_info_ import BBOX, GRID
+from phenotypic.tools_.measurement_info import BBOX, GRID
 
 
 class AutoGridFinder(GridFinder):
