+++ src/phenotypic/refine/_grid_oversized_object_remover.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from phenotypic.abc_ import GridObjectRefiner
 from phenotypic.tools_.constants_ import OBJECT
-from phenotypic.tools_.measurement_info_ import BBOX
+from phenotypic.tools_.measurement_info import BBOX
 
 
 class GridOversizedObjectRemover(GridObjectRefiner):
