+++ src/phenotypic/refine/_grid_section_largest.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
 
 from phenotypic.abc_ import ObjectRefiner
 from phenotypic.measure import MeasureSize
-from phenotypic.tools_.measurement_info_ import SIZE, GRID
+from phenotypic.tools_.measurement_info import SIZE, GRID
 from phenotypic.tools_.constants_ import OBJECT
