+++ src/phenotypic/refine/_merge_grid_obj.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
 
 from phenotypic.abc_ import GridObjectRefiner
 from phenotypic.tools_.constants_ import OBJECT
-from phenotypic.tools_.measurement_info_ import GRID
+from phenotypic.tools_.measurement_info import GRID
 
 
 class MergeGridObj(GridObjectRefiner):
