+++ src/phenotypic/refine/_min_residual_error_reducer.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from phenotypic.abc_ import GridObjectRefiner
 from phenotypic.measure import MeasureGridLinRegStats
-from phenotypic.tools_.measurement_info_ import GRID_LINREG_STATS
+from phenotypic.tools_.measurement_info import GRID_LINREG_STATS
 
 
 class ReduceMultipleGridObjects(GridObjectRefiner):
