+++ src/phenotypic/refine/_residual_outlier_remover.py
@@ -9,7 +9,7 @@ from typing import Optional
 
 from phenotypic.abc_ import GridObjectRefiner
 from phenotypic.measure import MeasureGridLinRegStats
-from phenotypic.tools_.measurement_info_ import GRID_LINREG_STATS, GRID
+from phenotypic.tools_.measurement_info import GRID_LINREG_STATS, GRID
 
 
 class ResidualOutlierRemover(GridObjectRefiner):
