+++ src/phenotypic/tools_/measurement_info/__init__.py
@@ -1,0 +1,53 @@
+"""Centralized measurement information enumerations for the PhenoTypic library.
+
+This subpackage contains all :class:`MeasurementInfo` subclasses used across
+the library, standardizing measurement naming conventions, metadata, and
+documentation. Each class lives in its own module (``_<name>.py``) and is
+re-exported from here, so the public import surface stays stable:
+
+    from phenotypic.tools_.measurement_info import BBOX, GRID, SHAPE
+"""
+
+from ._bbox import BBOX
+from ._color_composition import ColorComposition
+from ._color_hsv import ColorHSV
+from ._color_lab import ColorLab
+from ._color_xy import Colorxy
+from ._color_xyz import ColorXYZ
+from ._edge_correction import EDGE_CORRECTION
+from ._grid import GRID
+from ._grid_linreg_stats import GRID_LINREG_STATS
+from ._grid_spatial import GRID_SPATIAL
+from ._grid_spread import GRID_SPREAD
+from ._intensity import INTENSITY
+from ._linear_softplus_model import LINEAR_SOFTPLUS_MODEL
+from ._log_growth_model import LOG_GROWTH_MODEL
+from ._model_metrics import MODEL_METRICS
+from ._radial_expansion import RADIAL_EXPANSION
+from ._shape import SHAPE
+from ._size import SIZE
+from ._symmetric_zones import SYMMETRIC_ZONES
+from ._texture import TEXTURE
+
+__all__ = [
+    "BBOX",
+    "ColorComposition",
+    "ColorHSV",
+    "ColorLab",
+    "Colorxy",
+    "ColorXYZ",
+    "EDGE_CORRECTION",
+    "GRID",
+    "GRID_LINREG_STATS",
+    "GRID_SPATIAL",
+    "GRID_SPREAD",
+    "INTENSITY",
+    "LINEAR_SOFTPLUS_MODEL",
+    "LOG_GROWTH_MODEL",
+    "MODEL_METRICS",
+    "RADIAL_EXPANSION",
+    "SHAPE",
+    "SIZE",
+    "SYMMETRIC_ZONES",
+    "TEXTURE",
+]
