+++ src/phenotypic/tools_/measurement_info/_edge_correction.py
@@ -1,0 +1,18 @@
+"""Measurement info container for edge correction analysis results."""
+
+from phenotypic.abc_._measurement_info import MeasurementInfo
+
+
+class EDGE_CORRECTION(MeasurementInfo):
+    """Measurement info container for edge correction analysis results.
+
+    Provides metadata for measurement values produced by the EdgeCorrector,
+    organizing corrected colony measurements under the "EdgeCorrection" category.
+    """
+
+    @classmethod
+    def category(cls) -> str:
+        return "EdgeCorrection"
+
+    CORRECTED_CAP = "Cap", "The carrying capacity for the target measurement"
+    NEW_VAL = "NewVal", "The new value of the target measurement"
